--- sds-bindings-utils/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "sds-bindings-utils"
-version = "0.1.2"
-edition = "2024"
-license = "Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dd-sds = { path = "../sds" }
-
