--- sds-go/rust/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "dd_sds_go"
-version = "0.1.2"
-edition = "2024"
-license = "Apache-2.0"
-
-[lib]
-crate-type = ["staticlib", "cdylib", "lib"]
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_path_to_error = "0.1.17"
-dd-sds = { path = "../../sds" }
-sds-bindings-utils = { path = "../../sds-bindings-utils" }
-
-
