--- sds/tools/fuzz/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "sds-fuzz"
-version = "0.1.0"
-edition = "2024"
-
-[features]
-default = []
-hyperscan = ["dep:hyperscan"]
-manual_test = []
-
-[[bin]]
-name = "fuzz"
-path = "src/main.rs"
-
-[dependencies]
-afl = "0.14.5"
-ahash = "0.8.11"
-dd-sds = { path = "../../" }
-
-# optional
-hyperscan = { version = "0.3.2", features = ["static"], optional = true }
-rand = "0.8.5"
